requirements.txt
streamlit
pillow
opencv-python
rembg
onnxruntime
onnxruntime-gpu